Изменён код, запускающий фрагмент с информацией и списком репозиториев пользователя. Использован интерфейс Controller, фрагмент с информацией о пользователя теперь запускается в MainActivity.kt.
